Moved fonts to components as it's a react component
